DEV-66 Refactored DataService
